requirements.txt
Flask==1.1.1
Pillow==6.1.0
requests==2.22.0
ImageHash==4.0
python-whois==0.7.1
beautifulsoup4==4.8.0
